+++ tests/test_image_tags.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+"""This module has tests for the image_tags module."""
+
+# Copyright 2022-2023, United States Government as represented by the
+# Administrator of the National Aeronautics and Space Administration.
+# All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+# Reuse is permitted under the terms of the license.
+# The AUTHORS file and the LICENSE file are at the
+# top level of this library.
+
+import unittest
+
+from vipersci.vis.db import image_tags as it
+
+
+class TestImageTags(unittest.TestCase):
+    def test_init(self):
+        name = "Tag Name"
+        tag = it.ImageTag(name=name)
+        self.assertEqual(name, tag.name)